+++ tests/Rector/Dynamic/ArgumentReplacerRector/Test.php
@@ -1,0 +1,40 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Tests\Rector\Dynamic\ArgumentReplacerRector;
+
+use Rector\Rector\Dynamic\ArgumentReplacerRector;
+use Rector\Testing\PHPUnit\AbstractConfigurableRectorTestCase;
+
+final class Test extends AbstractConfigurableRectorTestCase
+{
+    public function test(): void
+    {
+//        $this->doTestFileMatchesExpectedContent(
+//            __DIR__ . '/wrong/wrong.php.inc',
+//            __DIR__ . '/correct/correct.php.inc'
+//        );
+//
+//        $this->doTestFileMatchesExpectedContent(
+//            __DIR__ . '/wrong/wrong2.php.inc',
+//            __DIR__ . '/correct/correct2.php.inc'
+//        );
+
+        $this->doTestFileMatchesExpectedContent(
+            __DIR__ . '/wrong/wrong3.php.inc',
+            __DIR__ . '/correct/correct3.php.inc'
+        );
+    }
+
+    protected function provideConfig(): string
+    {
+        return __DIR__ . '/config/rector.yml';
+    }
+
+    /**
+     * @return string[]
+     */
+    protected function getRectorClasses(): array
+    {
+        return [ArgumentReplacerRector::class];
+    }
+}
