+++ tests/Rector/Dynamic/ArgumentReplacerRector/correct/correct.php.inc
