+++ tests/Rector/Dynamic/ArgumentReplacerRector/correct/correct2.php.inc
