+++ tests/Rector/Dynamic/ArgumentReplacerRector/correct/correct3.php.inc
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+
+$container = new ContainerBuilder();
+$container->register('foo', 'stdClass')
+    ->setScope(false);
