+++ tests/Rector/Dynamic/ArgumentReplacerRector/wrong/wrong.php.inc
