+++ tests/Rector/Dynamic/ArgumentReplacerRector/wrong/wrong2.php.inc
