--- tests/Rector/Dynamic/MethodArgumentChangerRector/Test.php
@@ -1,35 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Tests\Rector\Dynamic\MethodArgumentChangerRector;
-
-use Rector\Rector\Dynamic\MethodArgumentChangerRector;
-use Rector\Testing\PHPUnit\AbstractConfigurableRectorTestCase;
-
-final class Test extends AbstractConfigurableRectorTestCase
-{
-    public function test(): void
-    {
-        $this->doTestFileMatchesExpectedContent(
-            __DIR__ . '/wrong/wrong.php.inc',
-            __DIR__ . '/correct/correct.php.inc'
-        );
-
-        $this->doTestFileMatchesExpectedContent(
-            __DIR__ . '/wrong/wrong2.php.inc',
-            __DIR__ . '/correct/correct2.php.inc'
-        );
-    }
-
-    protected function provideConfig(): string
-    {
-        return __DIR__ . '/config/rector.yml';
-    }
-
-    /**
-     * @return string[]
-     */
-    protected function getRectorClasses(): array
-    {
-        return [MethodArgumentChangerRector::class];
-    }
-}
